create the server atomicStack..make all operations atomicfor the atomic_stack

+++ runtime/gc/accounting/atomic_stack.h
@@ -270,7 +270,7 @@ class StructuredAtomicStack {
         stack_data_->capacity_ << ", to newCapacity: "<<  new_capacity;
     int32_t old_capacity;
     do {
-      old_capacity = stack_data_->capacity_;
+      old_capacity = static_cast<int32_t>(stack_data_->capacity_);
     } while (android_atomic_cas(old_capacity, static_cast<int32_t>(new_capacity),
         static_cast<volatile int*>(&stack_data_->capacity_)) != 0);
 //    stack_data_->capacity_ = new_capacity;
